// front/src/shared/api/generated/model/cartItemWithProduct.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { CartItemWithProductDescription } from './cartItemWithProductDescription';
import type { CartItemWithProductImageUrl } from './cartItemWithProductImageUrl';

export interface CartItemWithProduct {
  id: string;
  productId: string;
  userId: string;
  quantity: number;
  priceSnapshot: number;
  name: string;
  slug: string;
  description?: CartItemWithProductDescription;
  imageUrl?: CartItemWithProductImageUrl;
  categoryId: string;
}